# api/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AcceptInviteIn": {
        "additionalProperties": false,
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Full Name"
          },
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 200,
            "minLength": 10,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "AcceptInviteIn",
        "type": "object"
      },
...
      "CancelLinkIn": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 200,
            "minLength": 10,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "CancelLinkIn",
        "type": "object"
      },
      "ChangeIn": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangeIn",
        "type": "object"
      },
...
      "EmailChangeIn": {
        "additionalProperties": false,
        "properties": {
          "new_email": {
            "title": "New Email",
            "type": "string"
          }
        },
        "required": [
          "new_email"
        ],
        "title": "EmailChangeIn",
        "type": "object"
      },
      "EnrollConfirmIn": {
        "additionalProperties": false,
        "properties": {
          "challenge_token": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Challenge Token",
            "type": "string"
          },
          "code": {
            "maxLength": 12,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Label"
          }
        },
        "required": [
          "challenge_token",
          "code"
        ],
        "title": "EnrollConfirmIn",
        "type": "object"
      },
      "EnrollStartIn": {
        "additionalProperties": false,
        "properties": {
          "enrollment_token": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Enrollment Token"
          },
          "invite_context": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Invite Context"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Password"
          },
          "reauth": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reauth"
          },
          "replace": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Replace"
          }
        },
        "title": "EnrollStartIn",
        "type": "object"
      },
...
      "ForgotIn": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotIn",
        "type": "object"
      },
...
      "LoginIn": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Turnstile Token"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MfaRecoveryIn": {
        "additionalProperties": false,
        "properties": {
          "mfa_token": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Mfa Token",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "recovery_code": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Recovery Code",
            "type": "string"
          }
        },
        "required": [
          "mfa_token",
          "password",
          "recovery_code"
        ],
        "title": "MfaRecoveryIn",
        "type": "object"
      },
...
      "MfaVerifyIn": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 12,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "mfa_token": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Mfa Token",
            "type": "string"
          }
        },
        "required": [
          "mfa_token",
          "code"
        ],
        "title": "MfaVerifyIn",
        "type": "object"
      },
...
      "ResetIn": {
        "additionalProperties": false,
        "properties": {
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 200,
            "minLength": 10,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetIn",
        "type": "object"
      },
...
      "TokenIn": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 200,
            "minLength": 10,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TokenIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "P0 foundation (docs/architecture/08-api-design.md)",
    "title": "Veda Spaces API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/approvals/cancel-link": {
      "post": {
        "operationId": "approvals_cancel_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelLinkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Cancel a break-glass request via its signed link (06 \u00a77.5)",
        "tags": [
          "approvals"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "RBAC-021"
      }
    },
...
    "/api/v1/auth/.well-known/jwks.json": {
      "get": {
        "operationId": "auth_jwks_route",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.jwks_route",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-006",
        "x-requirement": "AUTH-004"
      }
    },
    "/api/v1/auth/email/cancel": {
      "post": {
        "operationId": "auth_cancel_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.cancel_email",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "USER-007"
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "auth_verify_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.verify_email",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "USER-007"
      }
    },
    "/api/v1/auth/invite/accept": {
      "post": {
        "operationId": "auth_accept_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.accept_invite",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "AUTH-011"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-001",
        "x-requirement": "AUTH-001"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-002",
        "x-requirement": "AUTH-007"
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "auth_logout_all",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke all own sessions",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-002",
        "x-requirement": "AUTH-007"
      }
    },
...
    "/api/v1/auth/me/email": {
      "put": {
        "operationId": "auth_change_own_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start the proposed-email workflow for yourself",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "USER-007"
      }
    },
...
    "/api/v1/auth/mfa/enroll/confirm": {
      "post": {
        "operationId": "auth_enroll_confirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.enroll_confirm",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-014"
      }
    },
    "/api/v1/auth/mfa/enroll/start": {
      "post": {
        "operationId": "auth_enroll_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.enroll_start",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-014"
      }
    },
    "/api/v1/auth/mfa/factor": {
      "delete": {
        "operationId": "auth_mfa_remove",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "auth.mfa_remove",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-003"
      }
    },
    "/api/v1/auth/mfa/recovery": {
      "post": {
        "operationId": "auth_mfa_recovery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaRecoveryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.mfa_recovery",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-001",
        "x-requirement": "MFA-013"
      }
    },
    "/api/v1/auth/mfa/recovery-codes": {
      "post": {
        "operationId": "auth_mfa_regenerate",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "auth.mfa_regenerate",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-005"
      }
    },
    "/api/v1/auth/mfa/step-up": {
      "post": {
        "operationId": "auth_mfa_step_up",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "auth.mfa_step_up",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-011"
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "auth_mfa_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.mfa_verify",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-001",
        "x-requirement": "MFA-001"
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "operationId": "auth_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "auth.change",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "AUTH-012"
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "operationId": "auth_forgot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.forgot",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "AUTH-008"
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "auth_reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "auth.reset",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-003",
        "x-requirement": "AUTH-008"
      }
    },
    "/api/v1/auth/reauth": {
      "post": {
        "operationId": "auth_reauth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "auth.reauth",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-004",
        "x-requirement": "MFA-011"
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "auth_refresh",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Rotate the refresh token",
        "tags": [
          "auth"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-002",
        "x-requirement": "AUTH-005"
      }
    },
...
    "/api/v1/public/leads": {
      "post": {
        "operationId": "leads_public_lead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicLeadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Problem"
          }
        },
        "summary": "Public website enquiry (anonymous, Turnstile-gated)",
        "tags": [
          "leads"
        ],
        "x-permission": null,
        "x-permission-any-of": null,
        "x-rbac-exception": "RBX-005",
        "x-requirement": "LEAD-001"
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.vedaspaces.com"
    }
  ]
}